pyproject.toml
[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517"]

[tool.pdm.version]
path = "eoapi/auth_utils/__init__.py"
source = "file"

[tool.pdm.build]
excludes = ["tests/", "**/.mypy_cache", "**/.DS_Store"]
includes = ["eoapi"]

[project]
authors = [
  {name = "Anthony Lukach", email = "anthony@developmentseed.org"},
]
dependencies = [
  "cryptography>=43.0.0",
  "fastapi-slim>=0.111.0",
  "pydantic-settings>=2.2.1",
  "pyjwt>=2.9.0",
]
description = "Authentication & authorization helpers for eoAPI"
dynamic = ["version"]
license = {file = "LICENSE"}
name = "eoapi.auth_utils"
readme = "README.md"
requires-python = ">=3.8"

[project.optional-dependencies]
lint = [
  "pre-commit",
]
testing = [
  "coverage",
  "httpx>=0.27.0",
  "jwcrypto>=1.5.6",
  "pytest>=6.0",
]
